pyproject.toml
[project]
name = "cyro"
version = "0.1.0"
description = "A terminal-based AI coding agent with dynamic subagent creation through markdown files"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai-slim[openai]>=0.4.9",
    "rich>=14.1.0",
    "typer>=0.16.0",
    "toml>=0.10.2",
    "pydantic-settings>=2.10.1",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "structlog>=25.4.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
cyro = "cyro:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "pyright>=1.1.403",
    "pytest>=8.0.0",
    "ruff>=0.12.5",
]
